Calculate directly standardised rates

Computes crude and directly standardised rates. Rates can be stratified by variables of interest.

dsr(
  data,
  event,
  time,
  age = "age_group",
  pop = "pop",
  strata = NULL,
  refdata = standardPopulation("esp2013"),
  mp = 1000,
  method = "normal",
  sig = 0.95,
  decimals = 4
)

Arguments

data

A data frame with counts and unit-times summarised by the standardisation variables.

event

Name of the variable in data that corresponds to the event counts.

time

Name of the variable in data that corresponds to the unit-time.

age

Name of the variable in data that corresponds to age

pop

Name of the variable in data that corresponds to population

strata

A variable within the input data frame for which rates are calculated by.

refdata

A data frame with population unit-times summarised by the standardisation variables. The unit-time variable name must named pop.

mp

A constant to multiply rates by (e.g. mp=1000 for rates per 1000).

method

Choose between normal, lognormal and gamma confidence intervals for crude and standardised rates. The default method is normal.

sig

The desired level of confidence in computing confidence intervals. The default is 0.95 for 95 percent CIs.

decimals

Round estimates to a desired decimal place.

Examples

#An example of calculating directly standardised rates
#Data from Table 1, Page 132 of Schoenbach (2000)

#State specific death counts and fu
df_study <- data.frame(state=rep(c('Miami',"Alaska"), c(5,5)),
                      age=rep(c('00-14','15-24','25-44','45-64','65+'),2),
                      deaths=c(136,57,208,1016,3605,59,18,37,90,81),
                      fu=c(114350,80259,133440,142670,92168,37164,20036,32693,14947,2077))

#US standard population
df_ref  <- data.frame(age=c('00-14','15-24','25-44','45-64','65+'),
                     pop=c(23961000,15420000,21353000,19601000,10685000))

#Directly standardised Rates (per 1000) - 95% CI's using the gamma method
my_results <- dsr(data = df_study,
                  event = "deaths",
                  time = "fu",
                  strata = "state",
                  age = "age",
                  refdata = df_ref)
#View results
my_results
#> # A tibble: 2 × 9
#>   state  Numerator Denominator `Crude Rate (per 1000)` `95% LCL (Crude)`
#>   <chr>      <dbl>       <dbl>                   <dbl>             <dbl>
#> 1 Miami       5022      562887                    8.92              8.68
#> 2 Alaska       285      106917                    2.67              2.36
#> # ℹ 4 more variables: `95% UCL (Crude)` <dbl>, `Std Rate (per 1000)` <dbl>,
#> #   `95% LCL (Std)` <dbl>, `95% UCL (Std)` <dbl>